VisitingCustomer.html
Code for HTML form to take the information of a customer visiting a departmental store such as name, contact phone no, preferred days of purchasing, favourite item (to be selected from a list of items), suggestions etc. One should provide button to Submit as well as Reset the form contents.